requirements.txt
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
moviepy
schedule
openai>=1.30.0
numpy>=1.24.0
pillow>=10.0.0
streamlit>=1.35.0
google-generativeai>=0.8.0
google-genai>=0.8.0
requests>=2.28.0
edge-tts>=6.1.0
